To find the solution for travelling salesman Problem
Algorithm:
Step1: Consider city 1 as the starting and ending point. Since the route is cyclic, we can      consider any point as a starting point.
Step2 :Generate all (n-1)! permutations of cities.
Step3:Calculate the cost of every permutation and keep track of the minimum cost permutation.
Step4: Return the permutation with minimum cost.